requirements.txt
# =============================================================================
# REVERSE TEMPLATING POC - Dependencias
# =============================================================================
# Instalar com: pip install -r requirements.txt
# =============================================================================

# Interface Web
streamlit>=1.28.0

# Leitura e extracao de PDF (texto nativo)
pdfplumber>=0.10.0

# Manipulacao de PDF (merge de paginas)
PyPDF2>=3.0.0

# Geracao de PDF (overlay)
reportlab>=4.0.0

# Preview de PDF e conversao para imagem
pymupdf>=1.23.0

# OCR - Tesseract (para PDFs escaneados)
pytesseract>=0.3.10
Pillow>=10.0.0

# Suporte a documentos Word
python-docx>=1.0.0

# LangChain para orquestracao de LLM
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0

# Google Generative AI (Gemini)
google-generativeai>=0.3.0

# Variaveis de ambiente
python-dotenv>=1.0.0

# Testes
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Banco de dados local
chromadb>=0.4.0

# =============================================================================
# INSTALACAO DO TESSERACT (necessario alem do pip):
# =============================================================================
# Windows:
#   1. Baixe: https://github.com/UB-Mannheim/tesseract/wiki
#   2. Instale e adicione ao PATH
#
# Linux (Ubuntu/Debian):
#   sudo apt-get install tesseract-ocr tesseract-ocr-por
#
# Mac:
#   brew install tesseract tesseract-lang
# =============================================================================
